+++ database/migrations/2024_02_12_212530_create_dpsm_akontacijes_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('SLOV_grupa_vrste_placanja')->nullable();
             $table->integer('sati')->nullable();
             $table->float('iznos',15,4)->nullable();
-            $table->integer('procenat')->nullable();
+            $table->float('procenat',15,4)->nullable();
             $table->string('POK2_obracun_minulog_rada')->nullable();
 
             $table->unsignedBigInteger('obracunski_koef_id');
